se agregan clases para inicialización de bases de datos

+++ src/main/java/com/decred/memories/payload/response/DecredMemoriesPriceResponse.java
@@ -1,17 +1,19 @@
 package com.decred.memories.payload.response;
 
+import java.math.BigDecimal;
+
 public class DecredMemoriesPriceResponse {
 
-	private Integer buyedTickets;
-	private Long ticketPrice;
+	private Long buyedTickets;
+	private BigDecimal ticketPrice;
 	private Long timeInSeconds;
 	private String dateForPrice;
 	private Integer window;
 	
 	public DecredMemoriesPriceResponse() {}
 
-	public DecredMemoriesPriceResponse(Integer buyedTickets, Long ticketPrice, Long timeInSeconds, String dateForPrice,
-			Integer window) {
+	public DecredMemoriesPriceResponse(Long buyedTickets, BigDecimal ticketPrice, Long timeInSeconds,
+			String dateForPrice, Integer window) {
 		super();
 		this.buyedTickets = buyedTickets;
 		this.ticketPrice = ticketPrice;
@@ -20,19 +22,19 @@ public class DecredMemoriesPriceResponse {
 		this.window = window;
 	}
 
-	public Integer getBuyedTickets() {
+	public Long getBuyedTickets() {
 		return buyedTickets;
 	}
 
-	public void setBuyedTickets(Integer buyedTickets) {
+	public void setBuyedTickets(Long buyedTickets) {
 		this.buyedTickets = buyedTickets;
 	}
 
-	public Long getTicketPrice() {
+	public BigDecimal getTicketPrice() {
 		return ticketPrice;
 	}
 
-	public void setTicketPrice(Long ticketPrice) {
+	public void setTicketPrice(BigDecimal ticketPrice) {
 		this.ticketPrice = ticketPrice;
 	}
